% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_lib.R
\name{design_combo_lib}
\alias{design_combo_lib}
\title{design a combinatorial library}
\usage{
design_combo_lib(
  design_file,
  gene_col = "Target Gene Symbol",
  guide_col = "sgRNA Sequence",
  guide_rank = "Pick Order",
  all_by_all_gene = F,
  row_genes = NULL,
  col_genes = NULL,
  ref_genes = NULL,
  gene_pairs = NULL,
  guide_pairing = "all",
  dual_orientation = F
)
}
\arguments{
\item{gene_pairs}{programmed gene pairs.}

\item{gene_list}{list of genes to design. Genes in this list will be designed
in an all by all fashion.}
}
\value{
combinatorial designs with columns gene1, gene2, guide1, guide2, dr
}
\description{
design a combinatorial library
}
