% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_lib.R
\name{get_combos}
\alias{get_combos}
\title{return all combinations of list, as a tibble}
\usage{
get_combos(ar, self = T)
}
\arguments{
\item{ar}{array of any type}

\item{self}{whether to include the same element twice}

\item{m}{number of elements to choose}

\item{reverse}{whether to flip the construct}
}
\value{
tibble with element1 and element2
}
\description{
return all combinations of list, as a tibble
}
