% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_lib.R
\name{shuffle_combo_lib}
\alias{shuffle_combo_lib}
\title{Shuffle the guide order of combinatorial library to avoid systematic biases
in guide order}
\usage{
shuffle_combo_lib(combo_lib, seed = NULL)
}
\value{
combinatorial designs with columns gene_1, gene_2, guide_1, guide_2
}
\description{
Shuffle the guide order of combinatorial library to avoid systematic biases
in guide order
}
